Qt 5.10.1 from debian buster

+++ src/widgets/dialogs/qfiledialog.cpp
@@ -43,7 +43,6 @@
 #include <private/qwidgetitemdata_p.h>
 #include "qfiledialog.h"
 
-#ifndef QT_NO_FILEDIALOG
 #include "qfiledialog_p.h"
 #include <private/qguiapplication_p.h>
 #include <qfontmetrics.h>
@@ -51,24 +50,19 @@
 #include <qheaderview.h>
 #include <qshortcut.h>
 #include <qgridlayout.h>
+#if QT_CONFIG(menu)
 #include <qmenu.h>
+#endif
+#if QT_CONFIG(messagebox)
 #include <qmessagebox.h>
-#include <qinputdialog.h>
+#endif
 #include <stdlib.h>
 #include <qsettings.h>
 #include <qdebug.h>
 #include <qmimedatabase.h>
 #include <qapplication.h>
 #include <qstylepainter.h>
-#if !defined(Q_OS_WINCE)
 #include "ui_qfiledialog.h"
-#else
-#define Q_EMBEDDED_SMALLSCREEN
-#include "ui_qfiledialog_embedded.h"
-#if defined(Q_OS_WINCE)
-extern bool qt_priv_ptr_valid;
-#endif
-#endif
 #if defined(Q_OS_UNIX)
 #include <pwd.h>
 #include <unistd.h> // for pathconf() on OS X
@@ -533,7 +527,7 @@ QFileDialogPrivate::QFileDialogPrivate()
         showHiddenAction(0),
         useDefaultCaption(true),
         qFileDialogUi(0),
-        options(new QFileDialogOptions)
+        options(QFileDialogOptions::create())
 {
 }
 
@@ -664,7 +658,7 @@ void QFileDialogPrivate::retranslateStrings()
     /* WIDGETS */
     if (options->useDefaultNameFilters())
         q->setNameFilter(QFileDialogOptions::defaultNameFilterString());
-    if (nativeDialogInUse)
+    if (!usingWidgets())
         return;
 
     QList<QAction*> actions = qFileDialogUi->treeView->header()->actions();
@@ -849,7 +843,7 @@ void QFileDialog::setVisible(bool visible)
             // Set WA_DontShowOnScreen so that QDialog::setVisible(visible) below
             // updates the state correctly, but skips showing the non-native version:
             setAttribute(Qt::WA_DontShowOnScreen);
-#ifndef QT_NO_FSCOMPLETER
+#if QT_CONFIG(fscompleter)
             // So the completer doesn't try to complete and therefore show a popup
             if (!d->nativeDialogInUse)
                 d->completer->setModel(0);
@@ -857,7 +851,7 @@ void QFileDialog::setVisible(bool visible)
         } else {
             d->createWidgets();
             setAttribute(Qt::WA_DontShowOnScreen, false);
-#ifndef QT_NO_FSCOMPLETER
+#if QT_CONFIG(fscompleter)
             if (!d->nativeDialogInUse) {
                 if (d->proxyModel != 0)
                     d->completer->setModel(d->proxyModel);
@@ -868,7 +862,7 @@ void QFileDialog::setVisible(bool visible)
         }
     }
 
-    if (d->usingWidgets())
+    if (visible && d->usingWidgets())
         d->qFileDialogUi->fileNameEdit->setFocus();
 
     QDialog::setVisible(visible);
@@ -900,6 +894,9 @@ void QFileDialogPrivate::_q_goToUrl(const QUrl &url)
         {QStandardPaths::standardLocations(QStandardPaths::PicturesLocation).last()},
         a native image picker dialog will be used for accessing the user's photo album.
         The filename returned can be loaded using QFile and related APIs.
+        For this to be enabled, the Info.plist assigned to QMAKE_INFO_PLIST in the
+        project file must contain the key \c NSPhotoLibraryUsageDescription. See
+        Info.plist documentation from Apple for more information regarding this key.
         This feature was added in Qt 5.5.
 */
 void QFileDialog::setDirectory(const QString &directory)
@@ -927,7 +924,7 @@ void QFileDialog::setDirectory(const QString &directory)
     if (!d->nativeDialogInUse) {
         d->qFileDialogUi->newFolderButton->setEnabled(d->model->flags(root) & Qt::ItemIsDropEnabled);
         if (root != d->rootIndex()) {
-#ifndef QT_NO_FSCOMPLETER
+#if QT_CONFIG(fscompleter)
             if (directory.endsWith(QLatin1Char('/')))
                 d->completer->setCompletionPrefix(newDirectory);
             else
@@ -1466,6 +1463,19 @@ void QFileDialog::selectNameFilter(const QString &filter)
 }
 
 /*!
+ * \since 5.9
+ * \return The mimetype of the file that the user selected in the file dialog.
+ */
+QString QFileDialog::selectedMimeTypeFilter() const
+{
+    Q_D(const QFileDialog);
+    if (!d->usingWidgets())
+        return d->selectedMimeTypeFilter_sys();
+
+    return d->options->initiallySelectedMimeTypeFilter();
+}
+
+/*!
     \since 4.4
 
     Returns the filter that the user selected in the file dialog.
@@ -1584,9 +1594,19 @@ QStringList QFileDialog::mimeTypeFilters() const
 */
 void QFileDialog::selectMimeTypeFilter(const QString &filter)
 {
-    const QString text = nameFilterForMime(filter);
-    if (!text.isEmpty())
-        selectNameFilter(text);
+    Q_D(QFileDialog);
+    d->options->setInitiallySelectedMimeTypeFilter(filter);
+
+    const QString filterForMime = nameFilterForMime(filter);
+
+    if (!d->usingWidgets()) {
+        d->selectMimeTypeFilter_sys(filter);
+        if (d->selectedMimeTypeFilter_sys().isEmpty() && !filterForMime.isEmpty()) {
+            selectNameFilter(filterForMime);
+        }
+    } else if (!filterForMime.isEmpty()) {
+        selectNameFilter(filterForMime);
+    }
 }
 
 #endif // QT_NO_MIMETYPE
@@ -1752,7 +1772,7 @@ int QFileDialogPrivate::maxNameLength(const QString &path)
 {
 #if defined(Q_OS_UNIX)
     return ::pathconf(QFile::encodeName(path).data(), _PC_NAME_MAX);
-#elif defined(Q_OS_WINCE) || defined(Q_OS_WINRT)
+#elif defined(Q_OS_WINRT)
     Q_UNUSED(path);
     return MAX_PATH;
 #elif defined(Q_OS_WIN)
@@ -2042,10 +2062,12 @@ QString QFileDialog::labelText(DialogLabel label) const
             button = d->qFileDialogUi->buttonBox->button(QDialogButtonBox::Save);
         if (button)
             return button->text();
+        break;
     case Reject:
         button = d->qFileDialogUi->buttonBox->button(QDialogButtonBox::Cancel);
         if (button)
             return button->text();
+        break;
     }
     return QString();
 }
@@ -2226,7 +2248,7 @@ QStringList QFileDialog::getOpenFileNames(QWidget *parent,
 }
 
 /*!
-    This is a convenience static function that will return or or more existing
+    This is a convenience static function that will return one or more existing
     files selected by the user. If the user presses Cancel, it returns an
     empty list.
 
@@ -2609,11 +2631,11 @@ void QFileDialog::accept()
         if (!info.exists())
             info = QFileInfo(d->getEnvironmentVariable(fn));
         if (!info.exists()) {
-#ifndef QT_NO_MESSAGEBOX
+#if QT_CONFIG(messagebox)
             QString message = tr("%1\nDirectory not found.\nPlease verify the "
                                           "correct directory name was given.");
             QMessageBox::warning(this, windowTitle(), message.arg(info.fileName()));
-#endif // QT_NO_MESSAGEBOX
+#endif // QT_CONFIG(messagebox)
             return;
         }
         if (info.isDir()) {
@@ -2641,7 +2663,7 @@ void QFileDialog::accept()
         if (!info.exists() || !confirmOverwrite() || acceptMode() == AcceptOpen) {
             d->emitFilesSelected(QStringList(fn));
             QDialog::accept();
-#ifndef QT_NO_MESSAGEBOX
+#if QT_CONFIG(messagebox)
         } else {
             if (QMessageBox::warning(this, windowTitle(),
                                      tr("%1 already exists.\nDo you want to replace it?")
@@ -2663,11 +2685,11 @@ void QFileDialog::accept()
             if (!info.exists())
                 info = QFileInfo(d->getEnvironmentVariable(file));
             if (!info.exists()) {
-#ifndef QT_NO_MESSAGEBOX
+#if QT_CONFIG(messagebox)
                 QString message = tr("%1\nFile not found.\nPlease verify the "
                                      "correct file name was given.");
                 QMessageBox::warning(this, windowTitle(), message.arg(info.fileName()));
-#endif // QT_NO_MESSAGEBOX
+#endif // QT_CONFIG(messagebox)
                 return;
             }
             if (info.isDir()) {
@@ -2889,10 +2911,10 @@ void QFileDialogPrivate::createWidgets()
 #ifndef QT_NO_SHORTCUT
     qFileDialogUi->fileNameLabel->setBuddy(qFileDialogUi->fileNameEdit);
 #endif
-#ifndef QT_NO_FSCOMPLETER
+#if QT_CONFIG(fscompleter)
     completer = new QFSCompleter(model, q);
     qFileDialogUi->fileNameEdit->setCompleter(completer);
-#endif // QT_NO_FSCOMPLETER
+#endif // QT_CONFIG(fscompleter)
 
     qFileDialogUi->fileNameEdit->setInputMethodHints(Qt::ImhNoPredictiveText);
 
@@ -3058,7 +3080,7 @@ void QFileDialog::setProxyModel(QAbstractProxyModel *proxyModel)
         proxyModel->setSourceModel(d->model);
         d->qFileDialogUi->listView->setModel(d->proxyModel);
         d->qFileDialogUi->treeView->setModel(d->proxyModel);
-#ifndef QT_NO_FSCOMPLETER
+#if QT_CONFIG(fscompleter)
         d->completer->setModel(d->proxyModel);
         d->completer->proxyModel = d->proxyModel;
 #endif
@@ -3068,7 +3090,7 @@ void QFileDialog::setProxyModel(QAbstractProxyModel *proxyModel)
         d->proxyModel = 0;
         d->qFileDialogUi->listView->setModel(d->model);
         d->qFileDialogUi->treeView->setModel(d->model);
-#ifndef QT_NO_FSCOMPLETER
+#if QT_CONFIG(fscompleter)
         d->completer->setModel(d->model);
         d->completer->sourceModel = d->model;
         d->completer->proxyModel = 0;
@@ -3333,7 +3355,7 @@ void QFileDialogPrivate::_q_showDetailsView()
 */
 void QFileDialogPrivate::_q_showContextMenu(const QPoint &position)
 {
-#ifdef QT_NO_MENU
+#if !QT_CONFIG(menu)
     Q_UNUSED(position);
 #else
     Q_Q(QFileDialog);
@@ -3362,7 +3384,7 @@ void QFileDialogPrivate::_q_showContextMenu(const QPoint &position)
         menu.addAction(newFolderAction);
     }
     menu.exec(view->viewport()->mapToGlobal(position));
-#endif // QT_NO_MENU
+#endif // QT_CONFIG(menu)
 }
 
 /*!
@@ -3397,7 +3419,7 @@ void QFileDialogPrivate::_q_deleteCurrent()
 
     QModelIndexList list = qFileDialogUi->listView->selectionModel()->selectedRows();
     for (int i = list.count() - 1; i >= 0; --i) {
-        QModelIndex index = list.at(i);
+        QPersistentModelIndex index = list.at(i);
         if (index == qFileDialogUi->listView->rootIndex())
             continue;
 
@@ -3407,10 +3429,9 @@ void QFileDialogPrivate::_q_deleteCurrent()
 
         QString fileName = index.data(QFileSystemModel::FileNameRole).toString();
         QString filePath = index.data(QFileSystemModel::FilePathRole).toString();
-        bool isDir = model->isDir(index);
 
         QFile::Permissions p(index.parent().data(QFileSystemModel::FilePermissions).toInt());
-#ifndef QT_NO_MESSAGEBOX
+#if QT_CONFIG(messagebox)
         Q_Q(QFileDialog);
         if (!(p & QFile::WriteUser) && (QMessageBox::warning(q_func(), QFileDialog::tr("Delete"),
                                     QFileDialog::tr("'%1' is write protected.\nDo you want to delete it anyway?")
@@ -3423,15 +3444,18 @@ void QFileDialogPrivate::_q_deleteCurrent()
                                       QMessageBox::Yes | QMessageBox::No, QMessageBox::No) == QMessageBox::No)
             return;
 
+        // the event loop has run, we have to validate if the index is valid because the model might have removed it.
+        if (!index.isValid())
+            return;
+
 #else
         if (!(p & QFile::WriteUser))
             return;
-#endif // QT_NO_MESSAGEBOX
+#endif // QT_CONFIG(messagebox)
 
-        // the event loop has run, we can NOT reuse index because the model might have removed it.
-        if (isDir) {
+        if (model->isDir(index) && !model->fileInfo(index).isSymLink()) {
             if (!removeDirectory(filePath)) {
-#ifndef QT_NO_MESSAGEBOX
+#if QT_CONFIG(messagebox)
             QMessageBox::warning(q, q->windowTitle(),
                                 QFileDialog::tr("Could not delete directory."));
 #endif
@@ -3612,7 +3636,7 @@ void QFileDialogPrivate::_q_enterDirectory(const QModelIndex &index)
 */
 void QFileDialogPrivate::_q_goToDirectory(const QString &path)
 {
- #ifndef QT_NO_MESSAGEBOX
+ #if QT_CONFIG(messagebox)
     Q_Q(QFileDialog);
 #endif
     QModelIndex index = qFileDialogUi->lookInCombo->model()->index(qFileDialogUi->lookInCombo->currentIndex(),
@@ -3631,12 +3655,12 @@ void QFileDialogPrivate::_q_goToDirectory(const QString &path)
 
     if (dir.exists() || path2.isEmpty() || path2 == model->myComputer().toString()) {
         _q_enterDirectory(index);
-#ifndef QT_NO_MESSAGEBOX
+#if QT_CONFIG(messagebox)
     } else {
         QString message = QFileDialog::tr("%1\nDirectory not found.\nPlease verify the "
                                           "correct directory name was given.");
         QMessageBox::warning(q, q->windowTitle(), message.arg(path2));
-#endif // QT_NO_MESSAGEBOX
+#endif // QT_CONFIG(messagebox)
     }
 }
 
@@ -3792,13 +3816,13 @@ void QFileDialogPrivate::_q_nativeEnterDirectory(const QUrl &directory)
 */
 bool QFileDialogPrivate::itemViewKeyboardEvent(QKeyEvent *event) {
 
+#if QT_CONFIG(shortcut)
     Q_Q(QFileDialog);
-
     if (event->matches(QKeySequence::Cancel)) {
         q->reject();
         return true;
     }
-
+#endif
     switch (event->key()) {
     case Qt::Key_Backspace:
         _q_navigateToParent();
@@ -3998,13 +4022,17 @@ void QFileDialogLineEdit::keyPressEvent(QKeyEvent *e)
     }
 #endif // QT_KEYPAD_NAVIGATION
 
+#if QT_CONFIG(shortcut)
     int key = e->key();
+#endif
     QLineEdit::keyPressEvent(e);
+#if QT_CONFIG(shortcut)
     if (!e->matches(QKeySequence::Cancel) && key != Qt::Key_Back)
+#endif
         e->accept();
 }
 
-#ifndef QT_NO_FSCOMPLETER
+#if QT_CONFIG(fscompleter)
 
 QString QFSCompleter::pathFromIndex(const QModelIndex &index) const
 {
@@ -4016,7 +4044,7 @@ QString QFSCompleter::pathFromIndex(const QModelIndex &index) const
     QString currentLocation = dirModel->rootPath();
     QString path = index.data(QFileSystemModel::FilePathRole).toString();
     if (!currentLocation.isEmpty() && path.startsWith(currentLocation)) {
-#if defined(Q_OS_UNIX) || defined(Q_OS_WINCE)
+#if defined(Q_OS_UNIX)
         if (currentLocation == QDir::separator())
             return path.mid(currentLocation.length());
 #endif
@@ -4104,11 +4132,9 @@ QStringList QFSCompleter::splitPath(const QString &path) const
     return parts;
 }
 
-#endif // QT_NO_COMPLETER
+#endif // QT_CONFIG(completer)
 
 
 QT_END_NAMESPACE
 
 #include "moc_qfiledialog.cpp"
-
-#endif // QT_NO_FILEDIALOG
